move upgrade diff stats to separate file
